A poorly made hash table for understanding the concept
This isn't a very good hash table and ideally I could have just use the
one that Java provides me with, but this was just an exercise. The
design is mostly bad because it will eat up a lot more space than is
necessary and I didn't use a good hashing algorithm (I just used the
simplest one). I would normally use a linked list or something else
instead of the array structure, but it was just for learning purposes.

I also appended the chatbot to it so that I could take extremely long
lists of words, fill a hash table and then pull out random words to
print out. This could be expanded upon, but it would be far more optimal
to just use the hash table stuff Java gives you.